+++ lib/src/clojure_lsp/source_paths.clj
@@ -43,10 +43,12 @@
 (defn ^:private classpath->source-paths [^java.nio.file.Path root-path classpath]
   (let [source-paths (->> classpath
                           (remove shared/jar-file?)
-                          (map (comp #(.getAbsoluteFile ^File %) io/file))
-                          (filter (fn [^File file]
-                                    (.startsWith (.toPath file) root-path)))
-                          (mapv #(.getCanonicalPath ^File %)))]
+                          (map (fn [path]
+                                 (if (shared/absolute-path? path)
+                                   path
+                                   (shared/get-canonical-path (io/file (str root-path) path)))))
+                          (filter (fn [^String abs-path]
+                                    (.startsWith abs-path (str root-path)))))]
     (when (seq source-paths)
       {:origins #{:classpath}
        :source-paths source-paths
@@ -60,15 +62,18 @@
         {:source-paths default-source-paths
          :origins #{:default}})))
 
+(defn ^:private absolutize-source-paths [source-paths root-path source-paths-ignore-regex]
+  (->> source-paths
+       set
+       (remove (fn [source-path]
+                 (let [relative-source-path (shared/relativize-filepath source-path (str root-path))]
+                   (some #(re-matches (re-pattern %) relative-source-path) source-paths-ignore-regex))))
+       (mapv #(->> % (shared/to-file root-path) .getCanonicalPath str))))
+
 (defn process-source-paths [settings root-path classpath given-source-paths]
   (let [source-paths-ignore-regex (get settings :source-paths-ignore-regex ["resources.*" "target.*"])
         {:keys [origins source-paths]} (resolve-source-paths root-path classpath given-source-paths)
-        final-source-paths (->> source-paths
-                                set
-                                (remove (fn [source-path]
-                                          (let [relative-source-path (shared/relativize-filepath source-path (str root-path))]
-                                            (some #(re-matches (re-pattern %) relative-source-path) source-paths-ignore-regex))))
-                                (mapv #(->> % (shared/to-file root-path) .getCanonicalPath str)))]
+        final-source-paths (absolutize-source-paths source-paths root-path source-paths-ignore-regex)]
     (when (contains? origins :settings) (logger/info startup-paths-logger-tag "Using given source-paths:" final-source-paths))
     (when (contains? origins :classpath) (logger/info startup-paths-logger-tag "Using source-paths from classpath:" final-source-paths))
     (when (contains? origins :default) (logger/info startup-paths-logger-tag "Using default source-paths:" final-source-paths))